backend/requirement.txt
python-dotenv==1.1.0
pinecone==6.0.2
openai==1.72.0
pymongo==4.12.0
redis==5.2.1

# Linting
ruff==0.11.4